requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
ConnPlotter==0.7a0
cycler==0.11.0
Cython==0.29.28
elephant==0.11.1
execnet==1.9.0
fonttools==4.32.0
future==0.18.2
idna==3.3
imagen==1.0
iniconfig==1.1.1
interval==1.0.0
junitparser==2.4.3
kiwisolver==1.4.2
lazyarray==0.5.2
matplotlib==3.5.1
mozaik==0.1.0
mpi4py==3.1.3
neo==0.10.2
nest-simulator==3.1
numpy==1.26.4
packaging==21.3
param==1.5.1
parameters==0.2.1
Pillow==9.1.0
pluggy==1.0.0
psutil==5.9.0
py==1.11.0
PyNN==0.10.0
pyparsing==3.0.8
pytest==7.1.1
pytest-forked==1.4.0
pytest-timeout==2.1.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
quantities==0.13.0
requests==2.27.1
scipy==1.8.0
six==1.16.0
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.9
